Feature

+++ components/File/UseFileActions.js
@@ -32,27 +32,41 @@ export const useFileActions = () => {
         const fileExtension = file.name.toLowerCase().split('.').pop();
         const imageSrc = getFileImage(fileExtension);
         const displayImageSrc = ['pdf', 'zip'].includes(fileExtension) ? imageSrc : file.imageUrl;
-
+    
         // Reference to the original file document
         const fileRef = doc(db, "files", file.id.toString());
-
+    
         // Reference to the 'deleted_files' collection
         const deletedFileRef = doc(collection(db, "deleted_files"));
-
-        // First, set the file document in the 'deleted_files' collection
-        await setDoc(deletedFileRef, {
-            ...file,
-            deletedAt: new Date(),
-        });
-
-        // Then delete the original file document
-        await deleteDoc(fileRef).then(() => {
-            setShowToastMsg('File Deleted!!!');
-            setTimeout(() => {
-                window.location.reload(); // Consider removing this if you manage state updates properly
-            }, 1500);
-            addNotification('image', { src: displayImageSrc, message: `${file.name} was deleted.` });
-        });
+    
+        // Fetch the file name
+        const fileSnapshot = await getDoc(fileRef);
+        if (fileSnapshot.exists()) {
+            const fileName = fileSnapshot.data().name;
+    
+            // First, set the file document in the 'deleted_files' collection
+            await setDoc(deletedFileRef, {
+                ...file,
+                deletedAt: new Date(),
+            });
+    
+            // Then delete the original file document
+            await deleteDoc(fileRef).then(() => {
+                setShowToastMsg('File Deleted!!!');
+                setTimeout(() => {
+                    window.location.reload(); // Consider removing this if you manage state updates properly
+                }, 1500);
+                
+                // Add a new field 'name' to the notification data
+                addNotification('image', {
+                    src: displayImageSrc,
+                    name: fileName, // Include the file name
+                    message: `${fileName} was deleted.`,
+                    isFile: true,
+                    isDeleted: true
+                });
+            });
+        }
     };
 
 
@@ -77,22 +91,38 @@ export const useFileActions = () => {
         const imageSrc = getFileImage(fileExtension);
         const displayImageSrc = ['pdf', 'zip'].includes(fileExtension) ? imageSrc : file.imageUrl;
         const newStarStatus = !starred;
+    
         const fileRef = doc(db, "files", file.id.toString());
+    
         await updateDoc(fileRef, {
             starred: newStarStatus,
         });
-
-        // Notify the user about the change
-        if (newStarStatus) {
-            addNotification('image', {
-                src: displayImageSrc,
-                message: `You've starred the file: ${file.name}.`
-            }, file.id);
+    
+        // Fetch the file name
+        const fileSnapshot = await getDoc(fileRef);
+        if (fileSnapshot.exists()) {
+            const fileName = fileSnapshot.data().name;
+    
+            // Notify the user about the change
+            if (newStarStatus) {
+                addNotification('image', {
+                    src: displayImageSrc,
+                    message: `You've starred the file: ${fileName}.`,
+                    isFile: true,
+                    isStarred: true,
+                    name: fileName, // Include the file name
+                });
+            } else {
+                addNotification('image', {
+                    src: displayImageSrc,
+                    message: `You've unstarred the file: ${fileName}.`,
+                    isFile: true,
+                    isUnstarred: true,
+                    name: fileName, // Include the file name
+                });
+            }
         } else {
-            addNotification('image', {
-                src: displayImageSrc,
-                message: `You've unstarred the file: ${file.name}.`
-            }, file.id);
+            console.error("Document does not exist!");
         }
     };
 
@@ -120,11 +150,15 @@ export const useFileActions = () => {
 
         // Clean up after the download
         document.body.removeChild(link);
+        const fileName = file.name;
 
         // Add a notification for the download action
         addNotification('image', {
             src: displayImageSrc,
-            message: `You've downloaded the file: ${file.name}.`
+            message: `You've downloaded the file: ${file.name}.`,
+            name: fileName,
+            isFile: true,
+            isDownloaded: true
         }, file.id);
     };
 
@@ -147,12 +181,16 @@ export const useFileActions = () => {
             window.location.reload(); // Consider removing this if you manage state updates properly
         }, 500);
     });
+    const fileName = newName
         setShowToastMsg('File renamed successfully!');
 
         // Add a notification for the renaming action
         addNotification('image', {
             src: displayImageSrc,
-            message: `File renamed to: ${newName}.`
+            message: `File renamed to: ${newName}.`,
+            name: fileName,
+            isFile: true,
+            isRename: true
         }, file.id);
     };
     
@@ -162,6 +200,14 @@ export const useFileActions = () => {
             sensitive: true,
         });
         setShowToastMsg('File locked successfully!');
+        // Add a notification for the renaming action
+        addNotification('image', {
+            src: displayImageSrc,
+            message: `File ${file.name} is locked now.`,
+            name: file.name,
+            isFile: true,
+            isLocked: true
+        }, file.id);
     };
 
     // UNLOCK FILE
@@ -171,6 +217,13 @@ export const useFileActions = () => {
             sensitive: false,
         });
         setShowToastMsg('File unlocked successfully!');
+        addNotification('image', {
+            src: displayImageSrc,
+            message: `File ${file.name} is unlocked now.`,
+            name: file.name,
+            isFile: true,
+            isUnlocked: true
+        }, file.id);
     };
 
     return { deleteFile, toggleStar, downloadFile, renameFile, lockFile, unlockFile };
